requirements.txt
pytest==7.2.2
PyHamcrest==2.0.4
requests==2.28.2
xmltodict==0.13.0